-돌봄 예약금 적립 구현체

+++ src/main/java/com/PetCare/service/Reservation/Reward/RewardServiceImpl.java
@@ -14,18 +14,18 @@ public class RewardServiceImpl implements RewardService {
     private final MemberRepository memberRepository;
 
     @Override
-    public void addReward(long customerId, double amount) {
+    public void addReward(long customerId, int amount) {
         Member customer = memberRepository.findById(customerId)
                 .orElseThrow(() -> new NoSuchElementException("고객 정보 조회에 실패했습니다."));
 
-        double reservationAmount = calculatorReward(amount);
+        int reservationAmount = calculatorReward(amount);
 
         customer.addRewardPoints(reservationAmount);
     }
 
     @Override
-    public double calculatorReward(double reservationAmount) { // 적립금 계산 로직 (예약금의 2% 적립)
-        return reservationAmount * 0.02;
+    public int calculatorReward(int reservationAmount) { // 적립금 계산 로직 (예약금의 2% 적립)
+        return (int) (reservationAmount * 0.02);
     }
 
 }
